==> dev/11-Colab-Combine-GPP-and-site-covariates.ipynb <==
# Getting started

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')

!pip install geopandas

import os
import geopandas as gpd
import glob
import pandas as pd
import numpy as np

# Combine covariate files in Google Cloud

!gsutil cp gs://bucketname/path-to-flux-tower-data-folder/GPPIN_AVG_ALL/*.csv .

file_list = []
for files in glob.glob("*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file, usecols = [1,2,3,4,5,6,7,8,9,10,11]) # Extract corresponding columns with covariates
  result['Site'] = file
  result_list.append(result)
result = pd.concat(result_list)

result.to_csv('final_extracted_GPP_covariates_all.csv')
!gsutil cp final_extracted_GPP_covariates_all.csv gs://bucketname/path-to-flux-tower-data-folder/"final_extracted_GPP_covariates_all.csv"

# Combine flux tower-derived GPP files in Google Cloud

!rm *.csv
!gsutil cp gs://bucketname/path-to-flux-tower-data-folder/GPP_sites_data/*.csv .

file_list = []
for files in glob.glob("*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file)
  result = pd.DataFrame(result, columns=['Date','GPP_PI_F_avg'])
  result['Site'] = file
  siteresult = pd.DataFrame(result)
  result_list.append(siteresult)
  result = pd.DataFrame(np.concatenate(result_list))

result.columns = ['Date', 'GPP_measured', 'Site']
result

result.to_csv('final_measured_GPP.csv')
!gsutil cp final_measured_GPP.csv gs://bucketname/path-to-flux-tower-data-folder/"final_measured_GPP"
