==> dev/12-Colab-Combine-NEE-and-site-covariates.ipynb <==
# Getting started

from google.colab import auth
auth.authenticate_user()

!earthengine authenticate
import ee
ee.Initialize()

!pip install geopandas

import os
import geopandas as gpd
import glob
import pandas as pd
import numpy as np

# Combine covariate files in Google Cloud

!gsutil cp gs://bucketname/path-to-flux-tower-data-folder/NEEINS_AVG_ALL/*.csv .

file_list = []
for files in glob.glob("*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file, usecols = [1,2,3,4,5,6,7,8,9,10,11,12], header = 0) # Extract corresponding columns with covariates
  result['Site'] = file
  siteresult = pd.DataFrame(result)
  result_list.append(siteresult)
  result = pd.DataFrame(np.concatenate(result_list))

result.columns = ['Clay','SWC1', 'SWC2', 'SW_IN', 'SW_IN_NLDAS', 'TA', 'TA_min', 'TS', 'VPD', 'date', 'fPAR', 'ppt', 'Site']
result

result.to_csv('final_extracted_NEE_covariates.csv')
!gsutil cp final_extracted_NEE_covariates.csv gs://bucketname/path-to-flux-tower-data-folder/"final_extracted_NEE_covariates_all.csv"

# Combine NEE files in Google Cloud

!rm *.csv
!gsutil cp gs://bucketname/path-to-flux-tower-data-folder/NEE_sites_data/*.csv .

file_list = []
for files in glob.glob("*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file)
  result = pd.DataFrame(result, columns=['Date','NEE_PI_F_avg'])
  result['Site'] = file
  siteresult = pd.DataFrame(result)
  result_list.append(siteresult)
  result = pd.DataFrame(np.concatenate(result_list))

result.columns = ['Date', 'NEE_measured', 'Site']
result

result.to_csv('final_measured_NEE.csv')
!gsutil cp final_measured_NEE.csv gs://bucketname/path-to-flux-tower-data-folder/"final_measured_NEE"
